Add files via upload
The .html files are a screenshot of a jupyter notebook of a linear regression analysis  of the province of Aceh in Indonesia. First on the list of 33. In my data there are 3 column, Biomass Loss, co2 Emissions, and Treecover Loss. I did a single regression analysis on each one and then a multiple regression analysis on the three. 
The Image is very useful for my hypothesis, as a contra indicator